requirements.txt
requests~=2.25.1
terminaltables~=3.1.0
python-dotenv~=0.17.1